requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
langchain